config.h.meson,platform_config: Remove unnecessary header checks
Autotools was keen on trying to make things compile for pre-ISO-C
compilers, so it checked for standard headers. There's no point in doing
that if we're going to assume a standard C++11 compiler. Let's just
remove them.

It also checked for various POSIX headers that we don't include
ourselves, so we can remove those too.