OutputStreamBuffer: Use std::vector rather than byte_string
std::vector was not specifically required to be contiguous in C++98,
despite that apparently being the intention. This code used the rather
horrible byte_string instead, because std::basic_string was required to
be contiguous.

C++03 clarified that std::vector must be contiguous. So we can just use
it instead and remove byte_string.