Add approval tests
mp3chop is so old that it didn't have very good tests. In the absence of
exhaustive tests, let's add a few approval tests. The expected output
files were generated using an older version of mp3chop and can be used
to show that the behaviour hasn't changed.

When we pass the mp3chop executable to our test script it is not passed
as a full path (see https://github.com/mesonbuild/meson/issues/2681 )
and the current directory won't be in our path so we need to take
special action to ensure that we can run it.